requirements.txt
Django>=3.2.0,<4.0.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
mediapipe>=0.8.10
face-recognition>=1.3.0 